BUG #28350: Stuff gets rounded in intraday if more than one queue is connnected to a skill
--HG--
branch : Release 7.5.390

+++ Domain/Forecasting/Statistic.cs
@@ -199,11 +199,10 @@ namespace Teleopti.Ccc.Domain.Forecasting
 
             if (statAnsweredTasks > 0)
             {
-                newStatisticTask.StatAverageAfterTaskTimeSeconds =
-                    (int)(statTaskAverageAfterTaskTimeSeconds / statAnsweredTasks);
-                newStatisticTask.StatAverageTaskTimeSeconds = (int)(statTaskAverageTaskTimeSeconds / statAnsweredTasks);
-                newStatisticTask.StatAverageQueueTimeSeconds = (int)(statTaskAverageQueueTimeSeconds / statAnsweredTasks);
-                newStatisticTask.StatAverageHandleTimeSeconds = (int)(statTaskAverageHandleTimeSeconds / statAnsweredTasks);
+                newStatisticTask.StatAverageAfterTaskTimeSeconds = statTaskAverageAfterTaskTimeSeconds / statAnsweredTasks;
+                newStatisticTask.StatAverageTaskTimeSeconds = statTaskAverageTaskTimeSeconds / statAnsweredTasks;
+                newStatisticTask.StatAverageQueueTimeSeconds = statTaskAverageQueueTimeSeconds / statAnsweredTasks;
+                newStatisticTask.StatAverageHandleTimeSeconds = statTaskAverageHandleTimeSeconds / statAnsweredTasks;
             }
             else
             {
